+++ src/Hacks/Level/Hitboxes/Hooks.hpp
@@ -4,6 +4,7 @@
 #include <Geode/modify/PlayLayer.hpp>
 #include <Geode/modify/GJBaseGameLayer.hpp>
 #include <Geode/modify/LevelEditorLayer.hpp>
+#include <Geode/modify/GJGameState.hpp>
 #include "HitboxNode.hpp"
 
 class $modify (HitboxBaseGameLayer, GJBaseGameLayer)
@@ -16,19 +17,22 @@ class $modify (HitboxBaseGameLayer, GJBaseGameLayer)
 
     void onTick(bool updateTrail);
 
+    // checkRepellPlayer runs right before updateCamera, so i use that to check if its the call at the end of a tick
+    void checkRepellPlayer();
+    void updateCamera(float dt);
+
     void resetLevelVariables();
     virtual bool init();
 };
 
 class $modify (HitboxPlayLayer, PlayLayer)
 {
-    virtual void postUpdate(float dt);
+    
 };
 
 class $modify (HitboxEditorLayer, LevelEditorLayer)
 {
     void onPlaytest();
 
-    virtual void postUpdate(float dt);
     virtual void updateVisibility(float dt);
 };