+++ src/Notifications/NotificationManager.cpp
@@ -22,6 +22,8 @@ NotificationManager* NotificationManager::get()
 
 void NotificationManager::notifyToast(std::string toastStr, float time)
 {
+    return;
+
     if (time == -1)
         time = NotificationsDuration::get()->getStringFloat();